Make acceptable PeopleMarkupCalculator more generic in nature.
The client may provide an object that responds to `.markup` and returns
the final markup percentage fraction to be applied. NuPack::Calculator does not
know how the fraction is calculated. The fraction calculator may or may
not depend upon the number of people the way it wants to or may not
depend at all.